adds property window

+++ whiskey2d/Whiskey2D/Core/Managers/Impl/DefaultLogManager.cs
@@ -92,6 +92,10 @@ namespace Whiskey2D.Core.Managers.Impl
         /// </summary>
         public void close()
         {
+            if (writer == null)
+                return;
+
+
             writer.Close();
 
             File.Delete(getOldLogPath());
